requirements.txt
aiohttp
uvloop
python-dotenv